+++ examples/DevDing/TemplateElement.cpp
@@ -1,13 +1,11 @@
 /**
  * @file TemplateElement.cpp
- * @brief Element Template class.
- *
  * @author Matthias Hertel, https://www.mathertel.de
  *
- * @Copyright Copyright (c) by Matthias Hertel, https://www.mathertel.de.
+ * @copyright Copyright (c) by Matthias Hertel, https://www.mathertel.de.
+ * This work is licensed under a BSD 3-Clause style license, see https://www.mathertel.de/License.aspx
  *
- * This work is licensed under a BSD 3-Clause style license,
- * https://www.mathertel.de/License.aspx.
+ * @brief Element Template class.
  *
  * More information on https://www.mathertel.de/Arduino
  *